requirements.txt
mysql-connector-python==8.0.32
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.0.3
MarkupSafe==2.1.2
Werkzeug==2.2.3
click==8.1.3
colorama==0.4.6
importlib-metadata==6.0.0
zipp==3.15.0